requirements.txt
aio-pika==8.3.0
aiohttp==3.8.3
aioredis==2.0.1
aiormq==6.6.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
charset-normalizer==2.1.1
cramjam==2.6.2
fastparquet==2022.12.0
frozenlist==1.3.3
fsspec==2022.11.0
gevent==22.10.2
greenlet==2.0.1
idna==3.4
multidict==6.0.4
numpy==1.24.1
orjson==3.8.4
packaging==23.0
pamqp==3.2.1
pandas==1.5.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
typing_extensions==4.4.0
websocket-client==1.4.2
websockets==10.4
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
